% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convert.scanonevar.to.empirical.ps.R
\name{convert.scanonevar.to.empirical.ps}
\alias{convert.scanonevar.to.empirical.ps}
\title{Convert Scanonevar from LODs to Empirical p-values}
\usage{
\method{convert}{scanonevar.to.empirical.ps}(scanonevar, perm.scan.maxes)
}
\arguments{
\item{scanonevar}{the \code{scanonevar} in LODs to be converted to emprirical p values}

\item{perm.scan.maxes}{the tbl_df object returned by scanonevar.perm, the maximum LOD score
   observed on a per-scan, per-chromosome-type basis in permutation scans.}
}
\value{
Returns a scanonevar object in terms of p-values, with \code{attr(x, 'units') = 'emp.ps'}.
}
\description{
\code{convert.scanonevar.to.empirical.ps} takes a scanonevar with LODs as units
   and maxes from permutation scans, estimates an extreme value distribution for the maxes,
   and returns the probability of observing the LOD scores in those EVDs.
}
\details{

}
\author{
Robert Corty \email{rcorty@gmail.com}
}
\seealso{
\code{\link{scanonevar}}, \code{\link{scanonevar.perm}}
}

