% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.additive.coef.from.2.genoprobs.R
\name{get.additive.coef.from.2.genoprobs}
\alias{get.additive.coef.from.2.genoprobs}
\title{Compute Additive Coefficient From Two Genotype Probabilities}
\usage{
get.additive.coef.from.2.genoprobs(pa, pb)
}
\arguments{
\item{pa}{A vector of length n giving the probability of A-type for each individual,
  or a matrix or data.frame of dimension n-by-2 giving probability of A-type and
  the probability of B-type for each individual.}

\item{pb}{A vector of length n giving the probability of B-type for each individual,
   or missing if \code{pa} is a data.frame or matrix.}
}
\value{
Additive coefficient vector.
}
\description{
\code{get.additive.coef.from.2.genoprobs} should not typically be called by a user.
   It is used to reliably set up model coefficients from genotype probabilities.
}
\details{

}
\author{
Robert Corty \email{rcorty@gmail.com}
}

