% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.genoprobs.by.marker.name.R
\name{get.genoprobs.by.marker.name}
\alias{get.genoprobs.by.marker.name}
\title{Get Genotype Probabilities From a Cross Object By Marker (or Psuedomarker) Name}
\usage{
get.genoprobs.by.marker.name(cross, marker.name)
}
\arguments{
\item{cross}{The cross from which the genetic information will be extracted.}

\item{marker.name}{The name of the marker where we want to know each individuals most likely genotype.}
}
\value{
Probability of each genotype at the given locus for each individual.
}
\description{
\code{get.genoprobs.by.marker.name} is an accessor function that returns the
   genotype probabilities of each individual in the cross at the given marker or pseudomarker.
}
\details{

}
\author{
Robert Corty \email{rcorty@gmail.com}
}

