% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.cross.w.genoprobs}
\alias{is.cross.w.genoprobs}
\title{is.cross.w.genoprobs}
\usage{
is.cross.w.genoprobs(x)
}
\arguments{
\item{x}{The object being tested for whether or not is is a cross object with genoprobs.}
}
\value{
TRUE if x is a cross object with valid genoprobs for each chromosome,
FALSE otherwise
}
\examples{
a <- qtl::sim.cross(map = qtl::sim.map())
is.cross.w.genoprobs(x = a)
b <- qtl::calc.genoprob(cross = a)
is.cross.w.genoprobs(x = b)

}
\author{
Robert W. Corty \email{rcorty@gmail.com}
}

