% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/is.scanonevar.R
\name{is.scanonevar}
\alias{is.scanonevar}
\title{is.scanonevar}
\usage{
is.scanonevar(sov)
}
\arguments{
\item{sov}{the object which is tested for being a scanonevar object}
}
\value{
Returns TRUE if 'sov' is a valid scanonevar object, with attribute 'why.not' an empty list
   Returns FALSE if 'sov' is not a valid scanonevar object with attribute 'why.not' a list of
   error messages for failed tests
}
\description{
Tests whether an object is a valid scanonevar object
}
\details{

}
\author{
Robert Corty \email{rcorty@gmail.com}
}
\seealso{
\code{\link{scanonevar}}, \code{\link{convert.scanonevar.to.empirical.ps}}
}

