% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAIGE_makeGroupFile.R
\name{makeGroupFileforRegions}
\alias{makeGroupFileforRegions}
\title{Run single variant or gene- or region-based score tests with SPA based on the linear/logistic mixed model.}
\usage{
makeGroupFileforRegions(
  bgenFile = "",
  bgenFileIndex = "",
  vcfFile = "",
  vcfFileIndex = "",
  vcfField = "DS",
  savFile = "",
  savFileIndex = "",
  bedFile = "",
  bimFile = "",
  famFile = "",
  AlleleOrder = "alt-first",
  regionFile = "",
  outputPrefix = ""
)
}
\arguments{
\item{bgenFile}{character. Path to bgen file. Currently version 1.2 with 8 bit compression is supported}

\item{bgenFileIndex}{character. Path to the .bgi file (index of the bgen file)}

\item{vcfFile}{character. Path to vcf file}

\item{vcfFileIndex}{character. Path to vcf index file. Indexed by tabix. Path to index for vcf file by tabix, .csi file using 'tabix --csi -p vcf file.vcf.gz'}

\item{vcfField}{character. genotype field in vcf file to use. "DS" for dosages or "GT" for genotypes. By default, "DS".}

\item{savFile}{character. Path to sav file}

\item{savFileIndex}{character. Path to index for sav file .s1r}

\item{bedFile}{character. Path to bed file (PLINK)}

\item{bimFile}{character. Path to bim file (PLINK)}

\item{famFile}{character. Path to fam file (PLINK)}

\item{AlleleOrder}{character. alt-first or ref-first for bgen or PLINK files. By default, alt-first}

\item{regionFile}{character. Path to a file containing genome regions to extract genetic markers for making group files}

\item{outputPrefix}{character. Path to the output files with prefix}

\item{sampleFile}{character. Path to the file that contains one column for IDs of samples in the bgen file. The file does not contain header lines.}
}
\value{
one or multiple group files, each for one chromosome
}
\description{
Run single variant or gene- or region-based score tests with SPA based on the linear/logistic mixed model.
}
