% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/margin.plot.R
\name{margin.plot}
\alias{margin.plot}
\title{Plot Phenotype of interest Averaged (Marginalized) Across Specified Markers and Phenotypes}
\usage{
margin.plot(cross, focal.phenotype.name, marginal.phen.names = NULL,
  marginal.marker.names = NULL, genotype.plotting.names = c("A", "H", "B"),
  ...)
}
\arguments{
\item{cross}{The cross object to be plotted}

\item{focal.phenotype.name}{the phenotype to put on the y-axis}

\item{marginal.phen.names}{a list of phenotypes to average over (put on the x-axis).}

\item{marginal.marker.names}{a list of marker names, whose values will be averaged over (put on the x-axis).}

\item{genotype.plotting.names}{Labels for the genotype groups.  Defaults to \code{c('AA', 'AB', 'BB')}.}
}
\value{
None.  Only makes plot.
}
\description{
\code{margin.plot} should be used to visually investigate the relationship
   between the phenotype of interest and other phenotypes.  \code{margin.plot} can also
   be used to visualize the relationship between the phenotype of interest and genetic
   loci of interest, but \code{predictive.plot} is usually preferrable.
}
\details{

}
\author{
Robert Corty \email{rcorty@gmail.com}
}

