% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.scanonevar.R
\name{plot.scanonevar}
\alias{plot.scanonevar}
\title{plot.scanonevar}
\usage{
\method{plot}{scanonevar}(scanonevar, chrs = unique(scanonevar$chr),
  col = c("black", "blue", "red", "darkgreen"), bandcol = "lightgray",
  legends = c("mean or var", "mean", "var"), legend.pos = "top", gap = 25,
  incl.markers = TRUE, main = attr(scanonevar, "pheno"), ylim = c(0, 1.05
  * max(coords.y.locus, na.rm = TRUE)), scanone.for.comparison = NULL,
  show.equations = length(chrs) != 1)
}
\arguments{
\item{scanonevar}{the \code{scanonevar} object to be plotted}

\item{chrs}{optionally, the subset of the chromosomes to plot}

\item{col}{optionally, a vector specifying the colors of the scan lines.  Defaults to \code{c("black", "blue", "red", "darkgreen")}.}

\item{bandcol}{optionally, a background color for the even-index chromosomes in this scan.}
}
\value{
Returns nothing.  Only makes the plot.
}
\description{
\code{plot.scanonevar} implements the plot generic for objects of class 'scanonevar'.
   Because scanonevar objects can be viewed in terms of LODs or empirical p-values,
   this plotting function checks the 'units' attribute to determine which to plot.
}
\details{


If such a strong signal was observed that the empirical p-value underflows R's
   float type, this function produces an error.  The author is open to suggestions on how
   to deal with this situation better.
}
\author{
Robert Corty \email{rcorty@gmail.com}
}
\seealso{
\code{\link{scanonevar}}, \code{\link{convert.scanonevar.to.empirical.ps}}
}

