% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predictive.plot.R
\name{predictive.plot}
\alias{predictive.plot}
\title{Plot Predictive Interval for Categorical Genotype/Phenotype Groups}
\usage{
predictive.plot(cross, mean.formula, var.formula, marker.name, phen.name,
  genotype.plotting.names = c("AA", "AB", "BB"))
}
\arguments{
\item{cross}{The cross object to be plotted}

\item{mean.formula}{The formula that describes the response, and the covariates and genetic effects that influence it.
  The left hand side of the ~ must be a single phenotype that is in the cross.
  The right hand side must use only phenotypes thata are in the cross, markers that are in the cross,
  and the special terms: mean.QTL.add (additive effect on the mean) and mean.QTL.dom (dominance deviation from additive on the mean).}

\item{var.formula}{The formula that describes the covariates and the genetic effects that influence residual (environmental) variation.
  There should be nothing on the left of the ~ (Inferred to be residual variation).
  The right hand side must use only phenotypes thata are in the cross, markers that are in the cross,
  and the special terms: var.QTL.add (additive effect on the variance) and var.QTL.dom (dominance deviation from additive on the variance).}

\item{marker.name}{The name of the marker the effects of which we want to investigate and visualize.}

\item{phen.name}{The categorical phenotype the effects of which we want to investigate and visualize.}

\item{genotype.plotting.names}{Labels for the genotype groups.  Defaults to \code{c('AA', 'AB', 'BB')}.}
}
\value{
None.  Only makes plot.
}
\description{
\code{predictive.plot} should be used to visually investigate loci identified
   with plot.scanonevar or summary.scanonevar.  The user can specify the same mean and variance
   formulae that were used in the scan, or specify new formulae to investigate interactions.
}
\details{

}
\author{
Robert Corty \email{rcorty@gmail.com}
}

