\docType{package}
\name{qtl-package}
\alias{qtl}
\alias{qtl-package}
\title{qtl: Tools for Analyzing QTL Experiments}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a set of tools to perform quantitative trait locus (QTL) analysis in experimental crosses. It is a reimplementation of the 'R/qtl' package to better handle high-dimensional data and complex cross designs. Broman et al. (2019) \doi{10.1534/genetics.118.301595}.
}
\section{Tutorials}{

\itemize{
\item \href{https://rqtl.org/tutorials/rqtltour.pdf}{A brief tour of R/qtl}
\item \href{https://rqtl.org/tutorials/rqtltour2.pdf}{A shorter tour of R/qtl}
\item \href{https://rqtl.org/tutorials/bcsft.pdf}{Users guide for new BCsFt tools for R/qtl}
\item \href{https://rqtl.org/tutorials/geneticmaps.pdf}{Genetic map construction with R/qtl}
\item \href{https://rqtl.org/tutorials/MQM-tour.pdf}{Tutorial on MQM within R/qtl}
}
}

\section{Google groups}{

\itemize{
\item For discussion/questions about R/qtl, join the \href{https://groups.google.com/forum/#!forum/rqtl-disc}{rqtl-disc google group}
\item Or join \href{https://groups.google.com/forum/#!forum/rqtl-announce}{rqtl-announce} for announcements about R/qtl and R/qtl2
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://rqtl.org}
  \item \url{https://github.com/kbroman/qtl}
  \item Report bugs at \url{https://github.com/kbroman/qtl/issues}
}

}
\author{
\strong{Maintainer}: Karl W Broman \email{broman@wisc.edu} (\href{https://orcid.org/0000-0002-4914-6671}{ORCID})

Authors:
\itemize{
  \item Karl W Broman \email{broman@wisc.edu} (\href{https://orcid.org/0000-0002-4914-6671}{ORCID})
  \item Hao Wu
  \item Gary Churchill (\href{https://orcid.org/0000-0001-9190-9284}{ORCID})
  \item Saunak Sen (\href{https://orcid.org/0000-0003-4519-6361}{ORCID})
  \item Danny Arents (\href{https://orcid.org/0000-0001-8738-0162}{ORCID})
  \item Robert Corty
  \item Timothee Flutre
  \item Ritsert Jansen
  \item Pjotr Prins (\href{https://orcid.org/0000-0002-8021-9162}{ORCID})
  \item Lars Ronnegard
  \item Rohan Shah
  \item Laura Shannon
  \item Quoc Tran
  \item Aaron Wolen
  \item Brian Yandell (\href{https://orcid.org/0000-0002-8774-9377}{ORCID})
}

Other contributors:
\itemize{
  \item R Core Team [contributor]
}

}
\keyword{internal}
