\name{qtlversion}
\alias{qtlversion}

\title{Installed version of R/qtl}

\description{
  Print the version number of the currently installed version of R/qtl.
}

\usage{qtlversion()}

%\arguments{
%}

\value{
  A character string with the version number of the currently installed
  version of R/qtl.
}

\author{Karl W Broman, \email{kbroman@biostat.wisc.edu} }

\examples{
  qtlversion()
}

\seealso{ \code{\link[base]{version}} }

\keyword{print}
