\name{summary.fitqtl}
\alias{summary.fitqtl}

\title{Summary of fit of qtl model}

\description{
  To be done.
}

\usage{
\method{summary}{fitqtl}(object, \dots)
}
\arguments{
 \item{object}{Output from \code{\link[qtl]{fitqtl}}.}
 \item{\dots}{Ignored at this point.}
}

\details{
  To be added.
}

\value{
  To be added.
}

\references{
  Sen, S. and Churchill, G. A. (2001) A statistical framework for quantitative
  trait mapping.  \emph{Genetics} \bold{159}, 371--387.
}

\author{Hao Wu, \email{hao@jax.org} }

\examples{
}

\seealso{ \code{\link[qtl]{fitqtl}}, \code{\link[qtl]{makeqtl}},
\code{\link[qtl]{scanqtl}} }  

\keyword{print}
