% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/validate.convert.scanonevar.to.emp.ps.R
\name{validate.convert.scanonevar.to.emp.ps}
\alias{validate.convert.scanonevar.to.emp.ps}
\title{Check the Compatibility of the Scanonevar to be Converted with the Permutations to be Used in the Conversion}
\usage{
validate.convert.scanonevar.to.emp.ps(scan, null.scan.maxes)
}
\arguments{
\item{scan}{the scanonevar to be converted}

\item{null.scan.maxes}{the maximum LODs observed in permutation (null) scans to be used in the conversion.}
}
\value{
Returns TRUE if the two arguments are compatible and FALSE otherwise.
}
\description{
\code{validate.convert.scanonevar.to.emp.ps} should not typically be called by a user.
 This function is used by \code{convert.scanonevar.to.emp.ps}
}
\details{

}
\author{
Robert Corty \email{rcorty@gmail.com}
}
\seealso{
\link{\code{convert.scanonevar.to.emp.ps}}, \code{\link{scanonevar}}, \code{\link{scanonevar.perm}}
}

